+++ examples/vanilla/main.js
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry-guardian/browser';
 
-const dsn = import.meta.env.VITE_DSN;
+const dsn = import.meta.env.VITE_DSN || 'http://localhost:3001/api/sentry/cmpzkjt2d0004hyyc79uh0x89';
 if (dsn) {
   Sentry.init({ dsn });
 }